refactor(BookItem): 아이템 영역을 컴포넌트로 분리
