+++ src/components/BookItem.tsx
@@ -1,0 +1,44 @@
+import type { BookDocument } from '../types';
+import { Button } from '@radix-ui/themes';
+import { ChevronDown } from 'lucide-react';
+import * as styles from './BookItem.css';
+import { Accordion } from 'radix-ui';
+
+const BookItem = ({ book, pageIndex }: { book: BookDocument; pageIndex: number }) => {
+  return (
+    <div className={styles.opener}>
+      <div className={styles.thumbnail}>
+        <img
+          src={book.thumbnail}
+          alt=""
+          width="48"
+          height="70"
+          loading={pageIndex ? 'lazy' : 'eager'}
+          decoding={pageIndex ? 'async' : 'sync'}
+        />
+      </div>
+      <div className={styles.title}>
+        <span className="title3">{book.title}</span>
+        <span className="body2 text-secondary">{book.authors}</span>
+      </div>
+      <div className={styles.price}>
+        <span className="title3">{book.sale_price}</span>
+      </div>
+      <div className={styles.buttons}>
+        <Button size="4" asChild>
+          <a href={book.url} target="_blank" rel="noopener noreferrer">
+            구매하기
+          </a>
+        </Button>
+        <Accordion.Trigger asChild>
+          <Button size="4" color="gray" variant="soft">
+            상세보기
+            <ChevronDown size={18} />
+          </Button>
+        </Accordion.Trigger>
+      </div>
+    </div>
+  );
+};
+
+export default BookItem;
