+++ src/components/index.ts
@@ -1,5 +1,6 @@
 export * from './shared';
 export { default as BookContext } from './BookContext';
+export { default as BookItem } from './BookItem';
 export { default as BookList } from './BookList';
 export { default as DetailSearch } from './DetailSearch';
 export { default as LikedList } from './LikedList';
